# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: BharatStudio Alerts API
  version: 1.0.0-draft
  description: >-
    Draft v1 REST contract for BharatStudio Alerts and Companion. This contract
    is not approved for production until L01 review and verification pass.
servers:
  - url: https://api.example.invalid
    description: Placeholder only; production domain is not yet approved.
...
paths:
...
  /v1/channels/{channelId}/test-alert:
    post:
      tags: [Configuration]
      operationId: createTestAlert
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestAlertRequest'
      responses:
        '202':
          description: Test alert accepted into the durable alert path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/channels/{channelId}/alert-history:
    get:
      tags: [History]
      operationId: listAlertHistory
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated alert/payment history visible to the role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertHistoryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/channels/{channelId}/moderation/{alertId}:
    post:
      tags: [History]
      operationId: moderateAlert
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/AlertId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationRequest'
      responses:
        '200':
          description: Moderation action recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/channels/{channelId}/billing:
    get:
      tags: [Billing]
      operationId: getBillingView
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Creator-safe billing and plan view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/channels/{channelId}/entitlements:
    get:
      tags: [Creator]
      operationId: getChannelEntitlements
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Server-resolved channel entitlements.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/channels/{channelId}/billing/subscription:
    post:
      tags: [Billing]
      operationId: createBillingSubscription
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBillingSubscriptionRequest'
      responses:
        '201':
          description: Server-created Razorpay subscription projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingSubscriptionResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/RetryableFailure'
...
components:
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
...
    AlertId:
      name: alertId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 256 }
      description: Opaque history cursor. Current form is ISO-8601 timestamp followed by "|" and the event UUID; timestamp-only legacy cursors remain readable during migration.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    viewerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-viewer-session-token
    overlayToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-overlay-token
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: OIDC-service-identity-token
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated identity lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    RetryableFailure:
      description: Temporary service failure; the client may retry.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Request conflicts with current resource/version state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [schemaVersion, errorCode, message, traceId]
      properties:
        schemaVersion: { type: string, const: v1 }
        errorCode: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        traceId: { type: string, minLength: 1, maxLength: 128 }
        retryable: { type: boolean }
...
    TestAlertRequest:
      type: object
      additionalProperties: false
      required: [displayName, message]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 500 }
        queueIds:
          type: array
          maxItems: 64
          items: { type: string, format: uuid }
    AlertAccepted:
      type: object
      additionalProperties: false
      required: [schemaVersion, eventId, traceId, status]
      properties:
        schemaVersion: { type: string, const: v1 }
        eventId: { type: string, format: uuid }
        traceId: { type: string, minLength: 1, maxLength: 128 }
        status: { type: string, enum: [accepted, held] }
    AlertHistoryPage:
      type: object
      additionalProperties: false
      required: [schemaVersion, items, nextCursor]
      properties:
        schemaVersion: { type: string, const: v1 }
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/HistoryItem' }
        nextCursor: { type: [string, 'null'], maxLength: 256, description: Opaque cursor in the form timestamp|event UUID. }
    HistoryItem:
      type: object
      additionalProperties: false
      required: [eventId, sourceType, status, createdAt]
      properties:
        eventId: { type: string, format: uuid }
        sourceType: { type: string, enum: [payment, manual, companion] }
        status: { type: string, enum: [accepted, held, displayed, acknowledged, failed, quarantined, suppressed] }
        createdAt: { type: string, format: date-time }
        grossAmountPaise: { type: [integer, 'null'], minimum: 0 }
        currency: { type: [string, 'null'], const: INR }
        displayName: { type: [string, 'null'], maxLength: 80 }
        message: { type: [string, 'null'], maxLength: 500 }
    ModerationRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action: { type: string, enum: [approve, hold, suppress, replay] }
        reason: { type: [string, 'null'], maxLength: 240 }
    ModerationResult:
      type: object
      additionalProperties: false
      required: [schemaVersion, eventId, action, appliedAt]
      properties:
        schemaVersion: { type: string, const: v1 }
        eventId: { type: string, format: uuid }
        action: { type: string, enum: [approve, hold, suppress, replay] }
        appliedAt: { type: string, format: date-time }
    BillingView:
      type: object
      additionalProperties: false
      required: [schemaVersion, channelId, tier, monthlyPricePaise, annualMonthsCharged, annualServiceMonths, renewalState, billingInterval, autoRenew, currentPeriodEndsAt, priceProtectedUntil, priceSource]
      properties:
        schemaVersion: { type: string, const: v1 }
        channelId: { type: string, format: uuid }
        tier: { type: string, enum: [free, pro, creator, studio] }
        monthlyPricePaise: { type: integer, minimum: 0 }
        annualMonthsCharged: { type: integer, const: 10 }
        annualServiceMonths: { type: integer, const: 12 }
        renewalState: { type: string, enum: [not_applicable, active, past_due, cancelled] }
        nextRenewalAt: { type: [string, 'null'], format: date-time }
        billingInterval: { type: string, enum: [monthly, annual] }
        autoRenew: { type: boolean }
        currentPeriodEndsAt: { type: [string, 'null'], format: date-time }
        priceProtectedUntil: { type: [string, 'null'], format: date-time }
        priceSource: { type: string, enum: [current, grandfathered] }
    CreateBillingSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [tier, billingInterval]
      properties:
        tier: { type: string, enum: [pro, creator, studio] }
        billingInterval: { type: string, enum: [monthly, annual] }
    BillingSubscriptionResult:
      type: object
      additionalProperties: false
      required: [schemaVersion, provider, status, subscriptionId, tier, billingInterval, monthlyPricePaise, annualChargePaise, annualMonthsCharged, annualServiceMonths, checkoutUrl]
      properties:
        schemaVersion: { type: string, const: v1 }
        provider: { type: string, const: razorpay }
        status: { type: string, enum: [linked, pending] }
        subscriptionId: { type: string }
        tier: { type: string, enum: [pro, creator, studio] }
        billingInterval: { type: string, enum: [monthly, annual] }
        monthlyPricePaise: { type: integer, minimum: 1 }
        annualChargePaise: { type: integer, minimum: 1 }
        annualMonthsCharged: { type: integer, const: 10 }
        annualServiceMonths: { type: integer, const: 12 }
        checkoutUrl: { type: [string, 'null'], format: uri }
...
    EntitlementResult:
      type: object
      additionalProperties: false
      required: [schemaVersion, channelId, tier, source, entitlementVersion, values]
      properties:
        schemaVersion: { type: string, const: v1 }
        channelId: { type: string, format: uuid }
        tier: { type: string, enum: [free, pro, creator, studio] }
        source: { type: string, enum: [individual_plan] }
        entitlementVersion: { type: integer, minimum: 1 }
        values: { type: object, additionalProperties: true }
...